refactor: 서비스 레이어 도메인 로직으로 수정
