+++ src/main/java/lotto/controller/LottoMachine.java
@@ -7,9 +7,9 @@ import lotto.domain.Lotto;
 import lotto.domain.TotalPrize;
 import lotto.dto.LottoResponseDto;
 import lotto.dto.TotalPrizeResponseDto;
-import lotto.service.Issue;
-import lotto.service.ReturnRate;
-import lotto.service.WinningLottery;
+import lotto.domain.Issue;
+import lotto.domain.ReturnRate;
+import lotto.domain.WinningLottery;
 import lotto.util.InputParser;
 import lotto.view.InputView;
 import lotto.view.OutputView;
